feat: pokemonEvolution & pokemonMoves, chargement plus rapide PokemonDetail
- Le chargement de la page PokemonDetail est maintenant beaucoup plus rapide.
- Les évolutions & les capacités ne sont plus attendues avant de charger la UI. De fait, la navigation est fluide.
- Enfin, les composants sont des composants serveur, ce qui utilise le SSR offert par Next.js à fond.
