+++ package.json
@@ -42,6 +42,7 @@
     "next": "^15.3.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-loader-spinner": "^6.1.6",
     "superjson": "^1.13.3",
     "zod": "^3.25.76"
   },
